2 / 7 / 2023
Ladders and Snake markers added to the React Block components - Red designated Snakes and Blue indicates ladders
Each token given own placement through CSS and individual token checkers.
End of game changes - Button action stops and changes to "game finished" - Scoreboard changes to say "game finished"
Refactored some of the code.2

Bug found and resolved:

Bug - If you click the button rapidly, the wait timer is overridden and the placement of the tokens isn't registered.2

Solution - Wait state added to app - When a player lands on a snake or a ladder, the game will move to "wait state"
until the sleep time has passed and the state is changed back to non-waiting.